+++ packages/smarthr-ui/src/components/FormGroup/index.ts
@@ -1,0 +1,1 @@
+export { FormGroup } from './FormGroup'
